fix wallet connect
